requirements/prod.txt
Django==2.2.5
django-grappelli==2.13.1
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
Pillow==6.1.0
django-rest-swagger==2.1.2
psycopg2==2.8.3
django-modeltranslation==0.13.3
django-cors-headers==3.1.0
django-mptt==0.10.0
PyYAML==5.1.2
uWSGI==2.0.18
